Added Property Listing

+++ restapi.py
@@ -17,7 +17,7 @@ import db
 import util
 import json
 
-
+PROPERTY_PATH = 'static/property_listings/'
 # signup_put_args = reqparse.RequestParser()
 # signup_put_args.add_argument("contact_number", type=str, help="Phone number of the user")
 
@@ -104,8 +104,19 @@ user_args.add_argument('address', type=str,help='Missing User address', required
 user_args.add_argument('latitude', type=str,help='Missing Google maps latitude coordinate', required=True)
 user_args.add_argument('longitude', type=str,help='Missing Google maps longitutde coordinate', required=True)
 
+
 user_args_put = reqparse.RequestParser()
 user_args_put.add_argument('userID', type=str, help='Missing User ID', required=True)
+user_args_put.add_argument('user_fname', type=str,help='Missing User firstname', required=True)
+user_args_put.add_argument('user_mname', type=str,help='Missing User middlename')
+user_args_put.add_argument('user_lname', type=str,help='Missing User lastname', required=True)
+user_args_put.add_argument('user_birthdate', type=str,help='Missing User birthdate', required=True)
+user_args_put.add_argument('address', type=str,help='Missing User address', required=True)
+user_args_put.add_argument('phone_number', type=str,help='Missing Phone number', required=True)
+
+user_args_patch = reqparse.RequestParser()
+user_args_patch.add_argument('sessionID', type=str, help='Missing Session ID', required=True)
+user_args_patch.add_argument('userID', type=str, help='Missing User ID', required=True)
 
 #TO BE MOVED!!!!
 
@@ -142,6 +153,7 @@ message_post.add_argument('sent_at', type=str, help='Missing Time of Sending',re
 
 #USER API CLASS
 #=======================================================================================
+
 class user(Resource):
     def get(self):
         userID = request.args.get('userID')
@@ -196,22 +208,65 @@ class user(Resource):
                 return {'message':'Error user creation'},400
 
     def delete(self):
+        sessionID = request.args.get('sessionID')
         userID = request.args.get('userID')
 
-        check_existing = db.check_existing_data('user', 'userID', userID)
-        if check_existing:
-            delete_user = db.delete_data('user', 'userID', userID)
-            if delete_user:
-                return {
-                    'message': f'User with UserID:{userID} is deleted'
-                }, 200
+
+        check_session = db.get_specific_data('session', ['sessionID','userID','status'], [sessionID,userID,'valid'])
+
+        if check_session:
+            check_existing = db.check_existing_data('user', 'userID', userID)
+            if check_existing:
+                delete_user = db.delete_data('user', 'userID', userID)
+                if delete_user:
+                    return {
+                        'message': f'deleted'
+                    }, 200
+                else:
+                    return{
+                        'message':f'Error deleting user with UserID:{userID}'
+                    }, 400
             else:
-                return{
-                    'message':f'Error deleting user with UserID:{userID}'
-                }, 400
+                return abort(400,'Cannot delete. User not found')
         else:
-            return abort(400,'Cannot delete. User not found')
-    
+            return abort(400, 'Session expired')
+        
+    def patch(self):
+        userInfo = user_args_patch.parse_args()
+
+        userJson = request.json
+        
+        fields = []
+        data = []
+        
+        for k,v in userJson.items():
+            if v is not None:
+                fields.append(k)
+                data.append(v)  
+
+        check_session = db.get_specific_data('session', ['sessionID','userID','status'], [userJson['sessionID'],userJson['userID'],'valid'])
+
+        if check_session:
+            check_existing = db.check_existing_data('user', 'userID', userJson['userID'])
+            if check_existing:
+                update_user = db.update_data('user', ['userID','user_status'], [userJson['userID'],'deactivated'])
+            
+                if update_user:
+                    return {
+                        'message': f"deactivated"
+                    }, 200
+                else:
+                    return{
+                        'message': f"Error updating user with userID:{userJson['userID']}"
+                    },400
+            else:
+                return abort(400,'Cannot delete. User not found')
+        else:
+            return abort(400, 'Session expired')
+
+
+
+
     def put(self):
         userInfo = user_args_put.parse_args()
 
@@ -241,7 +296,47 @@ class user(Resource):
                 },400
         else:
             return {'message':f'User with userID: {userJson["userID"]} does not exist'},400
+#CHANGE PASSWORD API CLASS
+#=======================================================================================   
+changepassword_args = reqparse.RequestParser()
+changepassword_args.add_argument('sessionID', type=str,help='Missing Session ID', required=True)
+changepassword_args.add_argument('userID', type=str,help='Missing User ID', required=True)
+changepassword_args.add_argument('user_password', type=str,help='Missing Password', required=True)
+class changepassword(Resource):
+    def post(self):
+        passwordInfo = changepassword_args.parse_args()
+
+        passwordJson = request.json
+        userID = None
+        fields = []
+        data = []        
+        for k,v in passwordJson.items():
+            if v is not None:
         
+                fields.append(k)
+                data.append(v)
+        check_session = db.get_specific_data('session', ['sessionID','userID','status'], [passwordJson['sessionID'],passwordJson['userID'],'valid'])
+
+        if check_session:
+            check_existing = db.check_existing_data('user', 'userID', passwordJson['userID'])
+            if check_existing:
+                update_user = db.update_data('user', ['userID','user_password','user_password_hashed'], [ passwordJson['userID'], passwordJson['user_password'], hashMD5(passwordJson['user_password'])])
+            
+                if update_user:
+                    return {
+                        'message': f"passwordchanged"
+                    }, 200
+                else:
+                    return{
+                        'message': f"Error updating user with userID:{passwordJson['userID']}"
+                    },400
+            else:
+                return abort(400,'Cannot delete. User not found')
+        else:
+            return abort(400, 'Session expired')
+    
+
+#=======================================================================================       
 
 class Leasing(Resource):
     def get(self):
@@ -328,8 +423,14 @@ class Leasing_Documents(Resource):
         leasingDocs = db.get_data('leasing_documents', 'leasingID', leasingID)
         
         if leasingDocs is not None:
+            # file_path = f"static\{leasingID}\{leasingDocs['leasing_doc_name']}"
             file_path = f"static\{leasingID}\{leasingDocs['leasing_doc_name']}"
-            return send_file(file_path, attachment_filename=os.path.basename(file_path), as_attachment=False)
+            a = '0b4df80577e23758b287b06a49e2e6ce'
+            b = 'b86a625efa1234ab92a33e4858957437_contract.pdf'
+            file_path = f"static\{a}\{b}"
+            return send_file(file_path)
+
+          
         else:
             return abort(400,'Contract not found')
     
@@ -441,18 +542,18 @@ class login(Resource):
 session_args = reqparse.RequestParser()
 session_args.add_argument('sessionID', type=str, help='Missing Session ID', required=True)
 session_args.add_argument('userID', type=str, help='Missing User ID', required=True)
-session_args.add_argument('ipAddress', type=str, help='Missing device IP Address', required=True)
 
 class session(Resource):
+    # This is where session is checked if it is still valid or not
     def get(self):
         sessionID = request.args.get('sessionID')
-        sessionInfo = db.get_data('session', 'sessionID', sessionID)
+        sessionInfo = db.get_specific_data('session', ['sessionID','status'], [sessionID,'valid'])
         if sessionInfo is not None:
             userJson = json.dumps(sessionInfo,default=str)
             print(userJson)
-            return userJson, 200
+            return {'message':'Session valid'}, 200
         else:
-            return abort(400,'Session not found')
+            return {'message':'Session expired'},200
     def post(self):
         sessionInfo = session_args.parse_args()
 
@@ -483,6 +584,8 @@ class session(Resource):
                     return {'message':'Error session creation'},400
             else:
                 return abort(400,'User does not exist')
+            
+    # This is where logout occurs. Client sends a put request containing the sessionID. Server updates the data in the database by setting its status to 'expired'
     def put(self):
         sessionInfo = session_args.parse_args()
 
@@ -494,9 +597,12 @@ class session(Resource):
                 fields.append(k)
                 data.append(v)
 
-        for i in range(len(data)):
-            if data[i] == 'open':
-                data[i] = 'closed'
+        fields.append('status')
+        data.append('expired')
+        fields.append('logoutTime')
+        data.append(str(datetime.now()))
+
+        
         
 
         check_existing= db.check_existing_data('session', 'sessionID', sessionJson['sessionID'])
@@ -505,9 +611,9 @@ class session(Resource):
             if check_existing_user:
                 update_data_bool = db.update_data('session',fields,data)
                 if update_data_bool:
-                    return {'message':'Success session update'},201
+                    return {'message':'Session expires'},201
                 else:
-                    return {'message':'Error session update'},400
+                    return {'message':'Error session expiration'},400
             else:
                 return abort(400,'User does not exist') 
         else:
@@ -639,3 +745,172 @@ class complaint(Resource):
 # ADMIN API CLASS | CRUD
     
 #=======================================================================================
+
+
+
+#=======================================================================================
+# PROPERTY LISTINGS API CLASS | CRUD
+    
+#=======================================================================================
+
+class property(Resource):
+    def get(self):
+        userID = request.args.get('userID')
+        sessionID = request.args.get('sessionID')
+        propertyID = request.args.get('propertyID')
+        if userID and sessionID:
+            check_user = db.check_existing_data('user','userID',userID)
+            check_session = db.check_existing_data('session','sessionID', sessionID)
+
+            if check_user and check_session:
+                if propertyID:
+                    data = db.get_specific_data('property', ['propertyID'],[propertyID])
+                    print(data)
+               
+                    data['images'] = []
+                    for filename in os.listdir(f'static/property_listings/{data["propertyID"]}/images/'):
+                        if filename.endswith('.jpg') or filename.endswith('.jpeg') or filename.endswith('.png'):
+                            # data['images'].append(str(filename))
+                                
+                            data['images'].append(filename)
+                    print(data)
+             
+                    
+                else:
+                    data = db.get_items('property', 'userID',userID)
+                    print(data)
+                    for each in data:
+                        each['images'] = []
+                        for filename in os.listdir(f'static/property_listings/{each["propertyID"]}/images/'):
+                            if filename.endswith('.jpg') or filename.endswith('.jpeg') or filename.endswith('.png'):
+                                # data['images'].append(str(filename))
+                                
+                                each['images'].append(filename)
+                        print(each)
+             
+                response = jsonify(data)
+                response.headers.add('Access-Control-Allow-Origin', '*')
+                return response
+        else:
+            return abort(404, "Incomplete request data")
+    def post(self):
+ 
+        address = request.form['address']
+        propertyLandSize = request.form['propertyLandSize']
+        propertyLandSizeUnit = request.form['propertyLandSizeUnit']
+        legalLandDescription = request.form['legalLandDescription']
+        price = request.form['price']
+        propertyType = request.form['propertyType']
+        moreDetails = request.form['moreDetails']
+        document = request.files.getlist('document')
+        images = request.files.getlist('images')
+        latitude = 0
+        longitude = 0
+        sessionID = request.form['sessionID']
+        userID = request.form['userID']
+        property_status = 'pending'
+        created_at = str(datetime.now())
+
+        fields = []
+        data = []
+
+        fields.append('propertyID')
+        fields.append('userID')
+        fields.append('property_status')
+        fields.append('address')
+        fields.append('latitude')
+        fields.append('longitude')
+        fields.append('property_description')
+        fields.append('land_description')
+        fields.append('size')
+        fields.append('unit_type')
+        fields.append('price')
+        fields.append('property_type')
+        fields.append('created_at')
+        propertyID = generateUUID(
+            str(userID) + "," + property_status  + "," + address  + "," + str(latitude)  + "," + str(longitude) + "," + 
+            moreDetails + "," + legalLandDescription + "," + str(propertyLandSize) + "," + propertyLandSizeUnit + "," + 
+            str(price) + "," + propertyType + "," + created_at)
+        
+        data.append(propertyID)
+        data.append(userID)
+        data.append(property_status)
+        data.append(address)
+        data.append(latitude)
+        data.append(longitude)
+        data.append(moreDetails)
+        data.append(legalLandDescription)
+        data.append(propertyLandSize)
+        data.append(propertyLandSizeUnit)
+        data.append(price)
+        data.append(propertyType)
+        data.append(created_at)
+
+
+        print(address)
+        print(propertyLandSize)
+        print(propertyLandSizeUnit)
+        print(legalLandDescription)
+        print(price)
+        print(propertyType)
+        print(moreDetails)
+        print(document)
+        print(images)
+        print(latitude)
+        print(longitude)
+        print(sessionID)
+        print(userID)
+        print(property_status)
+        print(created_at)
+        check_existing_property = db.check_existing_data('property', 'propertyID', propertyID)
+
+        if check_existing_property:
+            return {'message':f'Property with propertyID: {propertyID} already exist'},409
+        else:
+            
+            print(propertyLandSize)
+            print(propertyLandSizeUnit)
+            print(legalLandDescription)
+            print(price)
+            print(propertyType)
+            print(moreDetails)
+            insert_property = db.insert_data('property',fields,data)
+
+            # Create a directory for the property listings if it doesn't exist
+            if not os.path.exists(PROPERTY_PATH ):
+                os.makedirs(PROPERTY_PATH )
+            
+            #Checks the directory of property listing with propertyID. If not present, will create one.
+            if not os.path.exists(PROPERTY_PATH+"/"+propertyID ):
+                os.makedirs(PROPERTY_PATH+"/"+propertyID )
+            
+            #Checks the directory of property listing images with propertyID. If not present, will create one.
+            if not os.path.exists(PROPERTY_PATH+"/"+propertyID+"/"+"images" ):
+                os.makedirs(PROPERTY_PATH+"/"+propertyID+"/"+"images" )
+
+            #Checks the directory of property listing documents with propertyID. If not present, will create one.
+            if not os.path.exists(PROPERTY_PATH+"/"+propertyID+"/"+"documents" ):
+                os.makedirs(PROPERTY_PATH+"/"+propertyID+"/"+"documents" )
+            
+            
+
+            for image in images:
+                image.save(os.path.join(PROPERTY_PATH+"/"+propertyID+"/"+"images", image.filename))
+            for doc in document:
+                doc.save(os.path.join(PROPERTY_PATH+"/"+propertyID+"/"+"documents", doc.filename))
+
+            return {'message':'success'},200
+    def put(self):
+        pass
+    def delete(self):
+        pass
+
+class propertyimages(Resource):
+    def get(self,propertyID,image):
+        filename = f'static/property_listings/{propertyID}/images/{image}'
+        return send_file(filename, mimetype='image/jpeg')
+        
+        
+
+
+
